tmp work
- update tpch to use dictionary and bitshuffle
- switch to uint16 codewords in dict encoding

+++ src/kudu/cfile/string_dict_block.h
@@ -86,7 +86,7 @@ class StringDictBlockBuilder : public BlockBuilder {
   // They should NOT be clear in the Reset() method.
   StringPlainBlockBuilder dict_block_;
 
-  std::tr1::unordered_map<StringPiece, uint32_t, GoodFastHash<StringPiece> > dictionary_;
+  std::tr1::unordered_map<StringPiece, uint16_t, GoodFastHash<StringPiece> > dictionary_;
   // Memory to hold the actual content for strings in the dictionary_.
   //
   // The size of it should be bigger than the size limit for dictionary block
